requirements.txt
# Snowflake ML Platform Requirements
# Core ML and Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0

# Snowflake Dependencies
snowflake-snowpark-python>=1.11.0
snowflake-ml-python>=1.0.0
snowflake-connector-python>=3.5.0

# Dashboard and Visualization
streamlit>=1.28.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Environment and Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Data Processing and Utilities
pytz>=2023.0
python-dateutil>=2.8.2
typing-extensions>=4.7.0

# Security and Authentication
cryptography>=41.0.0
pyjwt>=2.8.0
pyopenssl>=23.0.0

# AWS Integration (for Snowflake stages)
boto3>=1.28.0
botocore>=1.31.0

# Development and Performance
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.25.0
watchdog>=3.0.0

# Optional: Enhanced Performance
# Uncomment if you want faster data processing
# pyarrow>=12.0.0
# fastparquet>=2023.0.0

# Optional: Advanced ML Features
# Uncomment if you need additional ML capabilities
# tensorflow>=2.13.0
# torch>=2.0.0
# transformers>=4.30.0